Util-LeonardoKey: merge with UrlUtil class
