+++ src/main/kotlin/util/kmttapi/UrlUtil.kt
@@ -24,6 +24,55 @@ object UrlUtil {
         }
     }
 
+    /**
+     * Get media id from Leonardo (media service). Returns `null` if url isn't valid
+     *
+     * @sample getIdOrNullSample
+     */
+    fun getMediaIdOrNull(url: String): String? {
+        val regex = Regex("""(?<=leonardo\.osnova\.io\/)\S+?[^\/](?=\/|${'$'}|\s+)""")
+        val search = regex.find(url)
+        return search?.value
+    }
+
+    private fun getIdOrNullSample() {
+        val validURL = "https://leonardo.osnova.io/118ce574-1507-5a19-92a0-6b06da18fdf3/"
+        UrlUtil.getMediaIdOrNull(validURL) // 118ce574-1507-5a19-92a0-6b06da18fdf3
+
+        val noSlashURL = "https://leonardo.osnova.io/118ce574-1507-5a19-92a0-6b06da18fdf3"
+        UrlUtil.getMediaIdOrNull(noSlashURL) // 118ce574-1507-5a19-92a0-6b06da18fdf3
+
+        val invalidURL = "https://osnova.io/118ce574-1507-5a19-92a0-6b06da18fdf3"
+        UrlUtil.getMediaIdOrNull(invalidURL) // null
+    }
+
+    /**
+     * Get media id from Leonardo (media service). Throws [IllegalArgumentException] if url isn't valid
+     *
+     * @sample getIdSample
+     */
+    fun getMediaId(url: String): String {
+        val value = this.getMediaIdOrNull(url)
+
+        requireNotNull(value) {
+            "$this is not a valid leonardo.osnova.io (media) url "
+        }
+
+        return value
+    }
+
+    private fun getIdSample() {
+        val validURL = "https://leonardo.osnova.io/118ce574-1507-5a19-92a0-6b06da18fdf3/"
+        UrlUtil.getMediaId(validURL) // 118ce574-1507-5a19-92a0-6b06da18fdf3
+
+        val noSlashURL = "https://leonardo.osnova.io/118ce574-1507-5a19-92a0-6b06da18fdf3"
+        UrlUtil.getMediaId(noSlashURL) // 118ce574-1507-5a19-92a0-6b06da18fdf3
+
+        val invalidURL = "https://osnova.io/118ce574-1507-5a19-92a0-6b06da18fdf3"
+        UrlUtil.getMediaId(invalidURL) // IllegalArgumentException
+    }
+
+
     fun getProfileID(url: String): Long {
         val id = SharedRegex.userIdRegex.find(url)
