+++ version.properties
@@ -1,6 +1,6 @@
 #Previous build version
-#Tue Sep 13 17:36:10 MSK 2022
-BUILD_NUMBER=832
+#Wed Sep 14 17:11:22 MSK 2022
+BUILD_NUMBER=834
 MAJOR_VERSION=2
 MINOR_VERSION=0
 PATCH_VERSION=0
